requirements.txt
django==5.1.4
mysqlclient==2.2.6
djangorestframework==3.15.2
